finishing up

+++ app/Http/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 use App\Post;
 use DB;
@@ -45,13 +46,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Handle file upload
+        if($request->hasFile('cover_image')){
+            //Get file name with the extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just file name
+            $fileName=pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just extension
+            $extension=$request->file('cover_image')->getClientOriginalExtension();
+            //file name to store
+            $fileNameToStore=$fileName.'_'.time().'.'.$extension;
+            //upload the image
+            $path=$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+            $fileNameToStore="noimage.jpg";
+        }
 
         //creating a post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image= $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -96,10 +116,27 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
-        //creating a post
+        //Handle file upload
+        if($request->hasFile('cover_image')){
+            //Get file name with the extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just file name
+            $fileName=pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get just extension
+            $extension=$request->file('cover_image')->getClientOriginalExtension();
+            //file name to store
+            $fileNameToStore=$fileName.'_'.time().'.'.$extension;
+            //upload the image
+            $path=$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+
+        //updating a post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image=$fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -113,7 +150,19 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
+
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image !== 'noimage.jpg'){
+            //Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
